Change apteryx refresh timeout return type
Returning an integer type refresh timeout value
causes corruption of the actual timeout value returned
in apteryx.c:call_uint64_callback().
Changing the return type to uint64_t appears to fix the issue.

+++ lua.c
@@ -565,7 +565,7 @@ lua_apteryx_unwatch (lua_State *L)
     return 1;
 }
 
-static int
+static uint64_t
 lua_do_refresh (const char *path, size_t ref)
 {
     int res = 0;
